feat: add data-testid attributes to enable UI testing
This commit adds `data-testid` attributes to the `<img>` element in
`FaceRecognition.jsx` and the `<input>` element in `ImageLinkForm.jsx`.
These attributes are used to provide unique identifiers for elements
in the UI, which makes it easier to target and interact with them
during UI testing. This improves the testability and maintainability
of the application.